pyproject.toml
[tool.poetry]
name = "windeval_catalog"
description = "data catalogs for windeval"
version = "2020.05.01.1"
authors = ["windeval developers <windeval@googlegroups.com>"]

include = [
    "src/windeval_catalog/windeval_data_dir/",
    "src/windeval_catalog/windeval_catalogs/"
]

[tool.poetry.dependencies]
python = "^3.7"
intake = ""
intake_xarray = ""


[tool.poetry.dev-dependencies]
pytest = "^5.4.1"
pytest-cov = "^2.8.1"
black = "19.3b0"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
